requirements.txt
flask
gunicorn
requests
pillow
cloudinary
python-dotenv